Fix pagination for page 1
**Issue**

When click on pagination and go to page 2, and then click on page 1 to go back,
a 404 page is rendered.

**Expected**

The homepage should be rendered.

**Fix Justification**

When the pagination number 1 should be rendered, render the <a> tag href
attribute to be site.baseurl + '/'; in other words, render the index.html page